pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "homebuyer"
version = "0.1.0"
description = "Berkeley CA home sales data collector and analyzer"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "click>=8.1",
    "pandas>=2.1",
    "geopandas>=0.14",
    "shapely>=2.0",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "jupyter>=1.0",
    "matplotlib>=3.8",
]

[project.scripts]
homebuyer = "homebuyer.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/homebuyer"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
